shell scripts show as changed even though there's no content different?
on my desktop. but I'm not seeing the same thing on my laptop. which is weird, because they're in the same relative folder. still, I'm guessing being on a different computer is somehow messing with the overall file hash that git uses. which is really annoying, if so. Anyway, committing now just to declutter my actual changes.