learned some things about colors, heatmap.2
and why it makes more sense to drop scale="row" in favor of scale="none" if I want to show true 0s. Still have to refactor to take advantage of the explicit breaks in heatmap.2, at least for normed values. And to find a color palette that will work for raw values. The biggest frustration is that I want to cluster schools based on normed values (methodological focus), then re-use that ordering with raw values (methodological output), to avoid over-weighting schools with < 5 total disses. Not sure why passing the rowInd from the former as Rowv for the latter isn't working; I must be misreading something in the docs...